feat: create task api request in separate file
