+++ src/api/board.api.ts
@@ -1,6 +1,5 @@
 import request from './request';
 import { Board, BoardItem } from '../types/Entities/Board';
-import { TaskItem } from '../types/Entities/Task';
 
 //возвращает массив досок
 export const getAllBoards = async () => {
@@ -12,27 +11,6 @@ export const getAllBoards = async () => {
 export const getSingleBoard = async (boardId: string) => {
   const response = await request.get<BoardItem>(`/boards/${boardId}`);
   return response.data;
-};
-
-//обнавить column у Task
-export const changeTask = async ({
-  task,
-  newColumnId,
-}: {
-  task: TaskItem;
-  newColumnId: string;
-}) => {
-  const data = {
-    title: task.title,
-    order: task.order,
-    description: task.description,
-    userId: task.userId,
-    boardId: task.boardId,
-    columnId: newColumnId,
-  };
-  await request.put(`/boards/${task.boardId}/columns/${task.columnId}/tasks/${task.id}`, data);
-
-  return await getSingleBoard(task.boardId || '');
 };
 
 //удаляет доску по ее id
