+++ src/store/reducers/actionCreators.ts
@@ -1,5 +1,6 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { getSingleBoard, getAllBoards, removeBoard, changeTask } from '../../api/board.api';
+import { getSingleBoard, getAllBoards, removeBoard } from '../../api/board.api';
+import { changeTaskColumn } from '../../api/task.api';
 import { removeColumn, putColumn } from '../../api/column.api';
 import {
   fetchLogin,
@@ -29,4 +30,4 @@ export const deleteColumn = createAsyncThunk('boards/deleteBoard', removeColumn)
 
 export const updateColumn = createAsyncThunk('boards/deleteBoard', putColumn);
 
-export const updateTask = createAsyncThunk('tasks/updateTask', changeTask);
+export const updateTaskColumn = createAsyncThunk('tasks/updateTask', changeTaskColumn);
